yksi/django/requirements.txt
boto
celery
django
django-allauth
django-celery-email
django-cleanup
django-crispy-forms
django-environ
django-filter
django-rest-auth
django-simple-captcha
djangorestframework
django-storages-redux
dj-database-url
easy-thumbnails
gunicorn
markdown
mysqlclient
psycopg2
tweepy
